perf!: impl `ExchangeMut`
This commit reimplements `Exchange` in terms of what changes need to be
made to an underlying type in order to exchange it (i.e. `ExchangeMut`).
This allows for performance improvements such as mutating `Vec`tors
in-place, preventing `clone`s of large structures, etc. And, when it is
convenient, this mutation can be obscured by calling `exchange` instead
of `exchange_mut` so that a user's code does not have to be peppered
with `let mut` (unless they want to).

Note that `ExchangeRef` was not implemented, because it seemed to me
that the cases where it would be useful (e.g. `Vec::new` deduping) were
removed by the reimplementation of `Exchange`.